requirements.txt
# ── Core Django framework ─────────────────────────────────────────────────────────
Django==5.0.14
djangorestframework==3.16.0
django-cors-headers==4.7.0

# ── Production server ────────────────────────────────────────────────────────────
gunicorn==23.0.0

# ── Database adapter ─────────────────────────────────────────────────────────────
psycopg2-binary==2.9.10

# ── Image support ────────────────────────────────────────────────────────────────
Pillow==11.1.0

# ── Static file serving ─────────────────────────────────────────────────────────
WhiteNoise==6.12.0

# ── Environment config ────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── JWT authentication ──────────────────────────────────────────────────────────
PyJWT==2.10.1
